RI-7682 Move Azure OAuth FE constants to constants/azure.ts
Drop the temporary apiEnums.ts file now that all the REST-surface
enums it mirrored come from apiClient. AzureAuthStatus and
AZURE_OAUTH_STORAGE_KEY are FE/BE-shared values transmitted via the
OAuth callback HTML + localStorage rather than HTTP, so they live in
a dedicated constants/azure.ts.

Refs RI-7682
