+++ redisinsight/ui/src/constants/azure.ts
@@ -1,0 +1,7 @@
+export enum AzureAuthStatus {
+  Processing = 'processing',
+  Succeed = 'succeed',
+  Failed = 'failed',
+}
+
+export const AZURE_OAUTH_STORAGE_KEY = 'ri_azure_oauth_result'
